refactor: Extra initService and handleRequest to requestHandler.js
